+++ internal/lookup/postgres/types.go
@@ -3,7 +3,7 @@ package postgres
 // GORM models for demonstration
 // (You can move these to a shared db package if you wish)
 type GormPath struct {
-	ID   uint      `gorm:"primaryKey"`
+	ID   uint64    `gorm:"primaryKey"`
 	Path string    `gorm:"uniqueIndex"`
 	URLs []GormURL `gorm:"foreignKey:PathID"`
 }
@@ -13,8 +13,8 @@ func (GormPath) TableName() string {
 }
 
 type GormURL struct {
-	ID     uint `gorm:"primaryKey"`
-	PathID uint
+	ID     uint64 `gorm:"primaryKey"`
+	PathID uint64
 	URL    string
 }
